Changed the cost returned by the obstacle cost function
By default the obstacle cost function returns the cost of the end
point of the local trajectory.

This obviously isn't very useful when the local trajectory is sampled
at points along the global planner and the cost is all the same along the path.

If a local trajectory cuts through areas of varying cost, the more important
return value is the max.

However, without an additional cost function this causes drone not to move as
a common case exists where numerous local trajectories have the same cost.